+++ packages/ui/src/Form/PhoneInput.stories.tsx
@@ -1,8 +1,8 @@
-import React, {CSSProperties} from 'react';
+import React, {CSSProperties, useState} from 'react';
 import {Story, Meta} from '@storybook/react';
+import {PhoneIcon} from '@eruditorgroup/profi-icons';
 
 import {PhoneInput, PhoneInputProps} from './index';
-import {PhoneIcon} from '@eruditorgroup/profi-icons';
 
 const containerStyles: CSSProperties = {
   marginBottom: '30px',
@@ -15,29 +15,40 @@ export default {
   component: PhoneInput,
 } as Meta;
 
-const Template: Story<PhoneInputProps> = ({...args}) => (
-  <div>
-    <link
-      rel="prefetch"
-      href="https://cdn.profi.ru/profi-front-web/profi-design-system/bare-phone-input-flags.png"
-    />
-    <div style={containerStyles}>
-      <PhoneInput
-        {...args}
-        style={itemStyles}
-        name="phoneNumber"
-        trailing={<PhoneIcon />}
+const PhoneInputStateful: React.FC<PhoneInputProps> = (props) => {
+  const [value, setValue] = useState('');
+  return <PhoneInput {...props} value={value} onChange={setValue} />;
+};
+
+const Template: Story<PhoneInputProps> = ({...args}) => {
+  return (
+    <div>
+      <link
+        rel="prefetch"
+        href="https://cdn.profi.ru/profi-front-web/profi-design-system/bare-phone-input-flags.png"
       />
-      <PhoneInput {...args} defaultCountryCode="by" style={itemStyles} />
-      <PhoneInput
-        {...args}
-        defaultCountryCode="ua"
-        style={itemStyles}
-        withFocusScroll
-      />
+      <div style={containerStyles}>
+        <PhoneInputStateful
+          {...args}
+          style={itemStyles}
+          name="phoneNumber"
+          trailing={<PhoneIcon />}
+        />
+        <PhoneInputStateful
+          {...args}
+          defaultCountryCode="by"
+          style={itemStyles}
+        />
+        <PhoneInputStateful
+          {...args}
+          defaultCountryCode="ua"
+          style={itemStyles}
+          withFocusScroll
+        />
+      </div>
     </div>
-  </div>
-);
+  );
+};
 
 export const Default = Template.bind({});
 Default.args = {};